Feat: Added lookup functions
resolves #1

Added a variety of functions for making lookups on the state machine,
including checking for a match, finding the first instance of a pattern,
finding all instances of a pattern, and matching the beginning of a
string.

All of these functions use const char* instead of std::string to prevent
excess allocations. We could possibly switch to std::string_view in the
future for more semantically correct c++ code